Refactor helpers into module
