+++ app.py
@@ -43,6 +43,23 @@ from database import (
 )
 
 from utils.sync import upsert_record
+from utils.helpers import (
+    _hex_to_rgb,
+    _is_dark_color,
+    _chart_style,
+    _try_read_csv,
+    _parse_shopify,
+    _parse_qbo,
+    _normalize_hubspot_source,
+    fetch_hubspot_traffic_data,
+    format_dt,
+    trend,
+    format_minutes,
+    heatmap_bg,
+    inject_globals,
+    fetch_resources,
+    log_error,
+)
 
 # default theme colors
 DEFAULT_THEME_PRIMARY = "#1976d2"
@@ -76,411 +93,6 @@ THEMES = {
         "text": "#e0e0e0",
     },
 }
-
-
-def _hex_to_rgb(value):
-    """Return (r, g, b) for a hex color string."""
-    value = value.lstrip("#")
-    if len(value) == 3:
-        value = "".join(c * 2 for c in value)
-    try:
-        r, g, b = [int(value[i : i + 2], 16) for i in (0, 2, 4)]
-    except Exception:
-        return 0, 0, 0
-    return r, g, b
-
-
-def _is_dark_color(value):
-    """Return True if the color is dark based on luminance."""
-    r, g, b = _hex_to_rgb(value)
-    luminance = 0.2126 * r + 0.7152 * g + 0.0722 * b
-    return luminance < 128
-
-
-def _chart_style():
-    bg = get_setting("theme_background", DEFAULT_THEME_BACKGROUND)
-    return "dark_background" if _is_dark_color(bg) else "default"
-
-
-def _try_read_csv(data: bytes, encodings=None):
-    """Try reading CSV bytes with a series of encodings."""
-    encodings = encodings or ["utf-8", "utf-8-sig", "utf-16", "latin-1", "cp1252"]
-    for enc in encodings:
-        try:
-            return pd.read_csv(BytesIO(data), encoding=enc)
-        except Exception:
-            continue
-    raise ValueError("Unsupported CSV encoding")
-
-
-def _parse_shopify(file_storage):
-    data = file_storage.read()
-    file_storage.seek(0)
-    # try csv first then excel
-    try:
-        df = _try_read_csv(data)
-    except Exception:
-        try:
-            df = pd.read_excel(BytesIO(data))
-        except Exception as exc:
-            raise ValueError("Could not parse Shopify file") from exc
-    required = {"Created at", "Lineitem sku"}
-    if not required.issubset(set(df.columns)):
-        raise ValueError("Invalid Shopify data")
-    cleaned = df[
-        [
-            "Created at",
-            "Lineitem sku",
-            "Lineitem name",
-            "Lineitem quantity",
-            "Lineitem price",
-        ]
-    ].copy()
-    cleaned["Total"] = pd.to_numeric(df["Lineitem price"], errors="coerce").fillna(
-        0
-    ) * pd.to_numeric(df["Lineitem quantity"], errors="coerce").fillna(0)
-    cleaned.columns = ["created_at", "sku", "description", "quantity", "price", "total"]
-    return cleaned
-
-
-def _parse_qbo(file_storage):
-    data = file_storage.read()
-    file_storage.seek(0)
-    try:
-        df = pd.read_excel(BytesIO(data), skiprows=4)
-    except Exception:
-        try:
-            df = _try_read_csv(data)
-        except Exception as exc:
-            raise ValueError("Could not parse QuickBooks file") from exc
-    expected = {"transaction_date", "product_service"}
-    if set(df.columns[:2]).intersection(expected) != expected:
-        # Set proper headers if they appear exactly as expected from QBO
-        try:
-            df.columns = [
-                "deleted_code",
-                "transaction_date",
-                "transaction_type",
-                "transaction_number",
-                "customer_name",
-                "line_description",
-                "quantity",
-                "sales_price",
-                "amount",
-                "balance",
-                "product_service",
-            ]
-        except ValueError as exc:
-            raise ValueError("Invalid QuickBooks data") from exc
-    df = df[df.get("transaction_date").notna()]
-    sku_col = next(
-        (c for c in df.columns if c.lower().replace(" ", "_") == "sku"),
-        None,
-    )
-    if sku_col:
-        df["sku"] = df[sku_col]
-    else:
-        df["sku"] = pd.NA
-    cleaned = df[
-        [
-            "transaction_date",
-            "sku",
-            "line_description",
-            "quantity",
-            "sales_price",
-            "amount",
-        ]
-    ].copy()
-    cleaned.columns = ["created_at", "sku", "description", "quantity", "price", "total"]
-    return cleaned
-
-
-def _fetch_shopify_api(domain, token, since=None, next_url=None):
-    """Return Shopify orders, line item rows and raw JSON data.
-
-    If ``since`` is provided, only orders updated at or after that ISO timestamp
-    are fetched. ``next_url`` can be used to fetch a single page for batching.
-    """
-    headers = {"X-Shopify-Access-Token": token}
-    if next_url:
-        url = next_url
-        params = None
-    else:
-        url = f"https://{domain}/admin/api/2023-07/orders.json"
-        params = {"status": "any", "limit": 250}
-        if since:
-            params["updated_at_min"] = since
-    orders = []
-    line_items = []
-    resp = requests.get(url, headers=headers, params=params, timeout=10)
-    resp.raise_for_status()
-    data = resp.json().get("orders", [])
-    orders.extend(data)
-    link = resp.headers.get("Link", "")
-    next_link = None
-    for part in link.split(","):
-        if 'rel="next"' in part:
-            start = part.find("<") + 1
-            end = part.find(">")
-            next_link = part[start:end]
-            break
-    rows = []
-    for order in orders:
-        created = order.get("created_at")
-        order_id = order.get("id")
-        for idx, line in enumerate(order.get("line_items", [])):
-            rows.append(
-                {
-                    "created_at": created,
-                    "sku": line.get("sku"),
-                    "description": line.get("name"),
-                    "quantity": line.get("quantity"),
-                    "price": line.get("price"),
-                    "total": (
-                        float(line.get("price", 0)) * float(line.get("quantity", 0))
-                    ),
-                }
-            )
-            line_items.append({"order_id": order_id, "line_num": idx, "data": line})
-    df = pd.DataFrame(
-        rows, columns=["created_at", "sku", "description", "quantity", "price", "total"]
-    )
-    return df, orders, line_items, next_link
-
-
-def _fetch_shopify_list(domain, token, endpoint, key, since=None, next_url=None):
-    """Return a list of Shopify objects for the given endpoint.
-
-    If ``since`` is provided, only objects updated after that ISO timestamp are
-    retrieved. ``next_url`` allows fetching a single page for batching.
-    """
-    headers = {"X-Shopify-Access-Token": token}
-    if next_url:
-        url = next_url
-        params = None
-    else:
-        url = f"https://{domain}/admin/api/2023-07/{endpoint}.json"
-        params = {"limit": 250}
-        if since:
-            params["updated_at_min"] = since
-    resp = requests.get(url, headers=headers, params=params, timeout=10)
-    resp.raise_for_status()
-    data = resp.json().get(key, [])
-    link = resp.headers.get("Link", "")
-    next_link = None
-    for part in link.split(","):
-        if 'rel="next"' in part:
-            start = part.find("<") + 1
-            end = part.find(">")
-            next_link = part[start:end]
-            break
-    return data, next_link
-
-
-def _refresh_qbo_access(client_id, client_secret, refresh_token):
-    """Return a new access token and refresh token for QuickBooks."""
-    auth = requests.auth.HTTPBasicAuth(client_id, client_secret)
-    data = {"grant_type": "refresh_token", "refresh_token": refresh_token}
-    resp = requests.post(
-        "https://oauth.platform.intuit.com/oauth2/v1/tokens/bearer",
-        auth=auth,
-        data=data,
-        timeout=10,
-    )
-    resp.raise_for_status()
-    tokens = resp.json()
-    return tokens.get("access_token"), tokens.get("refresh_token", refresh_token)
-
-
-def _qbo_api_url(realm_id, path, environment="prod"):
-    """Construct a QuickBooks API URL for the given environment."""
-    base = (
-        "https://sandbox-quickbooks.api.intuit.com"
-        if environment == "sandbox"
-        else "https://quickbooks.api.intuit.com"
-    )
-    return f"{base}/v3/company/{realm_id}/{path}"
-
-
-def _qbo_txn_lines(
-    headers, realm_id, doc_type, environment="prod", item_map=None, start_pos=1
-):
-    """Return line item dicts and raw documents for the given QBO document.
-
-    Parameters
-    ----------
-    headers : dict
-        HTTP headers with authorization for the API.
-    realm_id : str
-        QuickBooks realm/company ID.
-    doc_type : str
-        Document type, e.g. ``SalesReceipt`` or ``Invoice``.
-    environment : str, optional
-        API environment. Defaults to ``"prod"``.
-    item_map : dict, optional
-        Mapping of QuickBooks item IDs to their ``Item.Sku`` values.
-    """
-    url = _qbo_api_url(realm_id, "query", environment)
-    q = f"select Id, DocNumber, TxnDate, Line from {doc_type}"
-    if doc_type == "Invoice":
-        # QuickBooks query language sometimes requires numeric values
-        # to be quoted as strings for equality comparisons
-        q += " where Balance = '0'"
-    q += f" startposition {start_pos} maxresults 1000"
-    resp = requests.get(
-        url,
-        headers=headers,
-        params={"query": q, "minorversion": 65},
-        timeout=10,
-    )
-    resp.raise_for_status()
-    data = resp.json().get("QueryResponse", {})
-    txns = data.get(doc_type, [])
-    next_pos = start_pos + 1000 if len(txns) >= 1000 else None
-    rows = []
-    docs = []
-    line_items = []
-    for tx in txns:
-        docs.append(tx)
-        created_at = tx.get("TxnDate")
-        doc_id = str(tx.get("Id") or tx.get("DocNumber") or "")
-        for idx, line in enumerate(tx.get("Line", [])):
-            detail = line.get("SalesItemLineDetail")
-            if not detail:
-                continue
-            ref = detail.get("ItemRef") or {}
-            item_id = str(ref.get("value") or "")
-            sku = item_map.get(item_id) if item_map else None
-            if not sku:
-                # Skip lines that don't map to a known Item.Sku
-                continue
-            desc = line.get("Description") or ref.get("name")
-            qty = detail.get("Qty") or 1
-            price = detail.get("UnitPrice") or 0
-            total = line.get("Amount")
-            if total is None:
-                try:
-                    total = float(price) * float(qty)
-                except Exception:
-                    total = 0
-            rows.append(
-                {
-                    "created_at": created_at,
-                    "sku": sku,
-                    "description": desc,
-                    "quantity": qty,
-                    "price": price,
-                    "total": total,
-                    "doc_type": doc_type,
-                }
-            )
-            line_items.append({"doc_id": doc_id, "line_num": idx, "data": line})
-    return rows, docs, line_items, next_pos
-
-
-def _fetch_qbo_items(headers, realm_id, environment="prod"):
-    """Return DataFrame of QBO items and an item ID to SKU map."""
-    url = _qbo_api_url(realm_id, "query", environment)
-    q = "select Id, Name, Sku from Item"
-    resp = requests.get(
-        url,
-        headers=headers,
-        params={"query": q, "minorversion": 65},
-        timeout=10,
-    )
-    resp.raise_for_status()
-    data = resp.json().get("QueryResponse", {})
-    item_rows = []
-    item_map = {}
-    for item in data.get("Item", []):
-        sku = item.get("Sku") or item.get("Name")
-        item_id = str(item.get("Id") or "")
-        if sku:
-            item_rows.append({"sku": sku})
-            if item_id:
-                item_map[item_id] = sku
-    items_df = pd.DataFrame(item_rows, columns=["sku"])
-    return items_df, item_map
-
-
-def _fetch_qbo_list(headers, realm_id, entity, environment="prod"):
-    """Return a list of QBO objects for the given entity."""
-    url = _qbo_api_url(realm_id, "query", environment)
-    q = f"select * from {entity} startposition 1 maxresults 1000"
-    resp = requests.get(
-        url,
-        headers=headers,
-        params={"query": q, "minorversion": 65},
-        timeout=10,
-    )
-    resp.raise_for_status()
-    data = resp.json().get("QueryResponse", {})
-    return data.get(entity, [])
-
-
-def _fetch_qbo_api(
-    client_id,
-    client_secret,
-    refresh_token,
-    realm_id,
-    environment="prod",
-    doc_type="SalesReceipt",
-    start_pos=1,
-    item_map=None,
-    fetch_lists=False,
-):
-    """Return a single page of QBO transactions and related data."""
-
-    access_token, new_refresh = _refresh_qbo_access(
-        client_id, client_secret, refresh_token
-    )
-    headers = {
-        "Authorization": f"Bearer {access_token}",
-        "Accept": "application/json",
-    }
-
-    items_df = pd.DataFrame()
-    if item_map is None or fetch_lists:
-        items_df, item_map = _fetch_qbo_items(headers, realm_id, environment)
-
-    line_rows, raw_docs, raw_lines, next_pos = _qbo_txn_lines(
-        headers, realm_id, doc_type, environment, item_map, start_pos=start_pos
-    )
-    txn_df = pd.DataFrame(
-        line_rows,
-        columns=[
-            "created_at",
-            "sku",
-            "description",
-            "quantity",
-            "price",
-            "total",
-            "doc_type",
-        ],
-    )
-
-    customers = payments = products = invoices = []
-    if fetch_lists:
-        customers = _fetch_qbo_list(headers, realm_id, "Customer", environment)
-        payments = _fetch_qbo_list(headers, realm_id, "Payment", environment)
-        products = _fetch_qbo_list(headers, realm_id, "Item", environment)
-        invoices = _fetch_qbo_list(headers, realm_id, "Invoice", environment)
-
-    return (
-        txn_df,
-        items_df,
-        raw_docs,
-        raw_lines,
-        new_refresh,
-        headers,
-        item_map,
-        next_pos,
-        customers,
-        payments,
-        products,
-        invoices,
-    )
-
 
 app = Flask(__name__)
 app.secret_key = "secret"
@@ -566,298 +178,12 @@ HUBSPOT_SOURCE_MAP = {
     "PAID_SEARCHES": "Paid Search",
 }
 
-
-def _normalize_hubspot_source(src):
-    """Return standardized source name or ``None`` if unknown."""
-    if not isinstance(src, str):
-        return None
-    key = src.strip().replace(" ", "_").upper()
-    return HUBSPOT_SOURCE_MAP.get(key)
-
-
-def fetch_hubspot_traffic_data(
-    token,
-    start_year: int = 2021,
-    end_year: int = 2025,
-    *,
-    retries: int = 3,
-    cache_dir: str | None = None,
-):
-    """Return monthly HubSpot traffic analytics for a date range.
-
-    This helper downloads aggregated metrics for each traffic source from the
-    HubSpot Analytics API between ``start_year`` and ``end_year`` inclusive.
-    Returned rows contain ``sessions``, ``bounce_rate`` and ``avg_time_min``
-    converted from seconds. The result can be consumed as a ``pandas.DataFrame``
-    or iterated over as a list of dictionaries, e.g.::
-
-        {
-            "year": 2024,
-            "month": "Jan",
-            "source": "Organic Search",
-            "sessions": 61,
-            "bounce_rate": 0.48,
-            "avg_time_min": 2.5,
-        }
-
-    Parameters
-    ----------
-    token : str
-        Private app token for HubSpot.
-    start_year : int, optional
-        First year to retrieve (default ``2021``).
-    end_year : int, optional
-        Last year to retrieve (default ``2025``).
-    retries : int, optional
-        Maximum number of request attempts when rate limited.
-    cache_dir : str | None, optional
-        Directory to persist raw API responses for debugging.
-
-    Notes
-    -----
-    All traffic sources are retrieved and then normalized using
-    ``_normalize_hubspot_source``. Entries with unrecognized sources are
-    discarded.
-
-    Returns
-    -------
-    pandas.DataFrame
-        Traffic metrics grouped by month and source.
-    """
-
-    url = "https://api.hubapi.com/analytics/v2/reports/sources/monthly"
-    headers = {"Authorization": f"Bearer {token}"}
-    params = {
-        "start": f"{start_year:04d}0101",
-        "end": f"{end_year:04d}1231",
-    }
-
-    offset = None
-    rows = []
-    page = 0
-    while True:
-        if offset is not None:
-            params["offset"] = offset
-        attempt = 0
-        while True:
-            try:
-                resp = requests.get(url, headers=headers, params=params, timeout=15)
-                add_api_response("fetch_hubspot", resp.status_code, resp.text[:2000])
-                if resp.status_code == 429 and attempt < retries - 1:
-                    time.sleep(2**attempt)
-                    attempt += 1
-                    continue
-                resp.raise_for_status()
-            except Exception:
-                if attempt >= retries - 1:
-                    raise
-                time.sleep(2**attempt)
-                attempt += 1
-                continue
-            break
-
-        payload = resp.json()
-        if cache_dir:
-            try:
-                os.makedirs(cache_dir, exist_ok=True)
-                with open(
-                    os.path.join(cache_dir, f"hubspot_{page}.json"),
-                    "w",
-                    encoding="utf-8",
-                ) as fh:
-                    json.dump(payload, fh)
-            except Exception:
-                pass
-        page += 1
-        data = payload.get("results") or payload.get("data")
-        if data is None and "reports" in payload:
-            rep = payload.get("reports")
-            if isinstance(rep, list) and rep and isinstance(rep[0], dict):
-                data = rep[0].get("data", rep[0])
-        if not isinstance(data, dict):
-            data = {}
-        for period, details in data.items():
-            try:
-                dt = datetime.strptime(period, "%Y-%m")
-            except ValueError:
-                try:
-                    dt = datetime.strptime(period, "%Y-%m-%d")
-                except ValueError:
-                    continue
-            for entry in details:
-                raw = (
-                    entry.get("breakdown")
-                    or entry.get("source")
-                    or entry.get("sourceType")
-                )
-                src = _normalize_hubspot_source(raw)
-                if not src:
-                    continue
-                try:
-                    sessions = float(entry.get("sessions") or entry.get("visits") or 0)
-                except Exception:
-                    sessions = 0.0
-                try:
-                    bounce = float(entry.get("bounceRate"))
-                except Exception:
-                    bounce = 0.0
-                sec = entry.get("avgSessionLengthSeconds") or entry.get(
-                    "avgSessionLength"
-                )
-                try:
-                    avg = float(sec) / 60 if sec is not None else 0.0
-                except Exception:
-                    avg = 0.0
-                rows.append(
-                    {
-                        "year": dt.year,
-                        "month": MONTHS_ORDER[dt.month - 1],
-                        "month_num": dt.month,
-                        "source": src,
-                        "sessions": sessions,
-                        "bounce_rate": bounce,
-                        "avg_time_min": avg,
-                    }
-                )
-        if not payload.get("hasMore") and not payload.get("has_more"):
-            break
-        offset = payload.get("offset")
-        if offset is None:
-            break
-    return pd.DataFrame(
-        rows,
-        columns=[
-            "year",
-            "month",
-            "month_num",
-            "source",
-            "sessions",
-            "bounce_rate",
-            "avg_time_min",
-        ],
-    )
-
-
-TRAFFIC_METRIC_LABELS = {
-    "sessions": "Sessions",
-    "avg_time": "Avg Time (min)",
-    "bounce_rate": "Bounce Rate %",
-}
-
-
-def format_dt(value):
-    """Format ISO timestamp into 'mm/dd/yy - h:mma/pm'."""
-    try:
-        dt = datetime.fromisoformat(value)
-        time_str = dt.strftime("%I:%M%p").lstrip("0").lower()
-        date_str = dt.strftime("%m/%d/%y")
-        return f"{date_str} - {time_str}"
-    except Exception:
-        return value
-
-
 app.jinja_env.filters["format_dt"] = format_dt
-
-
-def trend(value, compare=None):
-    """Return HTML with colored arrow indicating trend.
-
-    If ``compare`` is provided, its numeric sign is used to determine the
-    arrow direction while ``value`` is displayed unchanged. This allows
-    showing the previous period's value with an arrow based on the change
-    from the current period.
-    """
-    try:
-        if value in ("-", None):
-            return value
-
-        sign_source = compare if compare is not None else value
-        sign_str = str(sign_source).strip()
-        if sign_str == "∞":
-            arrow = "▲"
-            color = "has-text-success"
-        else:
-            cleaned = (
-                sign_str.replace("$", "").replace(",", "").replace("%", "").strip()
-            )
-            num = float(cleaned)
-            if num > 0:
-                arrow = "▲"
-                color = "has-text-success"
-            elif num < 0:
-                arrow = "▼"
-                color = "has-text-danger"
-            else:
-                return value
-        inline = "#0f9d58" if color == "has-text-success" else "#d93025"
-        return Markup(
-            f"<span class='no-wrap {color}' style='color:{inline}'><span class='trend-arrow'>{arrow}</span> {value}</span>"
-        )
-    except Exception:
-        return value
-
-
 app.jinja_env.filters["trend"] = trend
-
-
-def format_minutes(value):
-    """Format minutes with two decimals."""
-    try:
-        return f"{float(value):.2f}"
-    except Exception:
-        return value
-
-
 app.jinja_env.filters["format_minutes"] = format_minutes
-
-
-def heatmap_bg(diff):
-    """Return style attribute for heatmap cell based on diff."""
-    try:
-        val = float(diff)
-    except Exception:
-        return ""
-    if val > 0:
-        color = "rgba(16,185,129,"
-    elif val < 0:
-        color = "rgba(239,68,68,"
-    else:
-        return ""
-    intensity = min(0.8, abs(val) / 100)
-    return f"background-color:{color}{intensity});"
-
-
 app.jinja_env.filters["heatmap"] = heatmap_bg
 
-
-@app.context_processor
-def inject_globals():
-    theme = {
-        "primary": get_setting("theme_primary", DEFAULT_THEME_PRIMARY),
-        "highlight": get_setting("theme_highlight", DEFAULT_THEME_HIGHLIGHT),
-        "background": get_setting("theme_background", DEFAULT_THEME_BACKGROUND),
-        "text": get_setting("theme_text", DEFAULT_THEME_TEXT),
-    }
-    return {
-        "app_name": get_setting("app_title", "ultrasuite"),
-        "theme": theme,
-        "MONTHS_ORDER": MONTHS_ORDER,
-    }
-
-
-def fetch_resources(uri, rel):
-    """Return local file path for xhtml2pdf resource URIs."""
-    if uri.startswith("/static"):
-        return os.path.join(app.root_path, uri.lstrip("/"))
-    return uri
-
-
-def log_error(message):
-    """Record a message in the application log."""
-    try:
-        add_log(message)
-    except Exception:
-        pass
+app.context_processor(inject_globals)
 
 
 @app.route("/favicon.ico")
